Move classes that works with plt-files into new directory
