+++ GUI/PLT/Stroke.cs
@@ -1,7 +1,7 @@
 ﻿using GUI.Colors;
 using GeneralComponents;
 
-namespace GUI {
+namespace GUI.PLT {
     class Stroke {
         public Point2D Start { get; }
         public Point2D End { get; }
